Line_AI/line-heroku/requirements.txt
line-bot-sdk
flask
tensorflow
matplotlib
numpy
keras
ipython
Pillow
